用户管理，jwt鉴权,用户登录

+++ config/conf_jwt.go
@@ -2,6 +2,6 @@ package config
 
 type Jwt struct {
 	SecretKey  string `json:"secret_key" yaml:"secret_key"`   // 密钥
-	ExpireTime int    `json:"expire_time" yaml:"expire_time"` // 过期时间(单位为time.hour)
+	ExpireTime int64  `json:"expire_time" yaml:"expire_time"` // 过期时间(单位为time.hour)
 	Issuer     string `json:"issuer" yaml:"issuer"`           // 颁发者
 }
